+++ libs/utils/electron/src/lib/remote-utils.ts
@@ -8,6 +8,15 @@ export type RemoteConfig = {
   enabled: boolean;
 };
 
+// Cache for remote config
+let cachedConfig: RemoteConfig | null = null;
+let cachedMtime: number | null = null;
+let hasLoggedConfig = false;
+let cacheCheckInterval: NodeJS.Timeout | null = null;
+
+// Check interval in milliseconds (1 minute)
+const CACHE_CHECK_INTERVAL_MS = 60_000;
+
 function getRemoteConfigPath(): string {
   return join(app.getPath('userData'), 'remote-config.json');
 }
@@ -30,22 +39,78 @@ async function ensureRemoteConfigExists(): Promise<void> {
   }
 }
 
+/**
+ * Start background interval to check if config file has changed.
+ * Clears the cache if mtime differs, so next getRemoteConfig() call will re-read.
+ */
+function startCacheInvalidationCheck(): void {
+  if (cacheCheckInterval) return; // Already running
+
+  cacheCheckInterval = setInterval(async () => {
+    if (cachedMtime === null) return; // No cached value to invalidate
+
+    try {
+      const configPath = getRemoteConfigPath();
+      const fileStat = await stat(configPath);
+      const currentMtime = fileStat.mtimeMs;
+
+      if (currentMtime !== cachedMtime) {
+        // File has changed, clear cache
+        cachedConfig = null;
+        cachedMtime = null;
+        hasLoggedConfig = false;
+        // eslint-disable-next-line no-console
+        console.log('Remote config file changed, cache invalidated');
+      }
+    } catch {
+      // File might have been deleted, clear cache
+      cachedConfig = null;
+      cachedMtime = null;
+      hasLoggedConfig = false;
+    }
+  }, CACHE_CHECK_INTERVAL_MS);
+}
+
 export async function getRemoteConfig(): Promise<RemoteConfig> {
+  // Return cached value if available
+  if (cachedConfig !== null) {
+    return cachedConfig;
+  }
+
   await ensureRemoteConfigExists();
   const configPath = getRemoteConfigPath();
-  const configContent = await readFile(configPath, 'utf-8');
+
+  // Read file and get mtime
+  const [configContent, fileStat] = await Promise.all([
+    readFile(configPath, 'utf-8'),
+    stat(configPath),
+  ]);
+
   const remoteConfig = JSON.parse(configContent) as RemoteConfig;
-  if (remoteConfig.enabled) {
-    // eslint-disable-next-line no-console
-    console.log(
-      'Remote password is',
-      remoteConfig.password,
-      'config is at',
-      configPath,
-    );
-  } else {
-    // eslint-disable-next-line no-console
-    console.log('Remote access is disabled. Config is at', configPath);
+
+  // Cache the config and mtime
+  cachedConfig = remoteConfig;
+  cachedMtime = fileStat.mtimeMs;
+
+  // Log only on first load or after cache invalidation
+  if (!hasLoggedConfig) {
+    hasLoggedConfig = true;
+    if (remoteConfig.enabled) {
+      // eslint-disable-next-line no-console
+      console.log(
+        'Remote password is',
+        remoteConfig.password,
+        'config is at',
+        configPath,
+      );
+    } else {
+      // eslint-disable-next-line no-console
+      console.log('Remote access is disabled. Config is at', configPath);
+    }
   }
+
+  // Start background check if not already running
+  startCacheInvalidationCheck();
+
   return remoteConfig;
 }
